workitem 9208.  As part of the fix, introduction of CompressionMethod enum.  Also, some tests.
git-svn-id: https://dotnetzip.svn.codeplex.com/svn@47027 4ef1bfeb-0a34-4d02-9305-240e5b07a168

+++ Zip Tests/WinZipAesTests.cs
@@ -577,7 +577,7 @@ namespace Ionic.Zip.Tests.WinZipAes
                 {
                     if (!e.IsDirectory)
                     {
-                        Assert.AreEqual<short>(0, e.CompressionMethod);
+                        Assert.AreEqual<short>(0, (short)e.CompressionMethod);
 
                         e.ExtractWithPassword("unpack", password);
